Route _xyzzy+/a._b.c.tsx does not nest correctly
In Remix's route module convention, the following three route
module filenames are interpreted as an ordinary route, a pathless
layout route, and another ordinary route, each containing the next:

```
a.tsx
a._b.tsx
a._b.c.tsx
```

However, with remix-flat-routes, this does not work as expected if
the modules are contained within a folder, like this:

```
_xyzzy+/a.tsx
_xyzzy+/a._b.tsx
_xyzzy+/a._b.c.tsx
```

Add a (currently failing) unit test to demonstrate this bug.

Note: update Jest to 28.x because that version introduced
`test.failing`.